backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
mysqlclient==2.2.4
django-cors-headers==4.6.0
python-decouple==3.8
